Add fake checkout command
All it does is autocomplete branches. There's an unreachable path in the
code, and I believe a ton of the commands are not getting autocompleted
because of it.

There's probably a more efficient way to make it do a git checkout, but
I'm not thinking about it at close to midnight. I guess I could also add
a mapping to do the branch checkout action from branch mode?